### 2022-02-23 #4195 @editorialbot
[PRE REVIEW]: inaFaceAnalyzer: a Python toolbox for large-scale face-based description of gender representation in media with limited gender, racial and age biases
**Submitting author:** @DavidDoukhan (<a href="http://orcid.org/0000-0002-1645-7334">David Doukhan</a>)
**Repository:** https://github.com/ina-foss/inaFaceAnalyzer
**Branch with paper.md** (empty if default branch):
**Version:** v0.5.6
