+++ paper.bib
@@ -86,7 +86,8 @@ howpublished={\url{https://github.com/rcmalli/keras-vggface}},
   number={14},
   pages={103--122},
   year={2018},
-  publisher={Netherlands Institute for Sound and Vision}
+  publisher={Netherlands Institute for Sound and Vision},
+  doi={10.18146/2213-0969.2018.jethc156}
 }
 
 @inproceedings{doukhan2018open,
@@ -95,7 +96,8 @@ howpublished={\url{https://github.com/rcmalli/keras-vggface}},
   booktitle={2018 IEEE International Conference on Acoustics, Speech and Signal Processing (ICASSP)},
   pages={5214--5218},
   year={2018},
-  organization={IEEE}
+  organization={IEEE},
+  doi={10.1109/ICASSP.2018.8461471}
 }
 
 
