### 2022-02-24 #4195 @DavidDoukhan
> **Submitting author:** @DavidDoukhan ([David Doukhan](http://orcid.org/0000-0002-1645-7334)) **Repository:** https://github.com/ina-foss/inaFaceAnalyzer **Branch with paper.md** (empty if default branch): **Version:** v0.5.6 **Editor:** Pending **Reviewers:** **Managing EiC:** Arfon Smith
>
> ## Status
> [![status](https://camo.githubusercontent.com/88c274ee4f22776bb4f68be5cece040940e381c0e369615ee9f65cdbd2680468/68747470733a2f2f6a6f73732e7468656f6a2e6f72672f7061706572732f34373038363135316637343666333231363563373238373039373863313339382f7374617475732e737667)](https://joss.theoj.org/papers/47086151f746f32165c72870978c1398)
>
> Status badge code:
>
> ```
> ```
>
> **Author instructions**
>
> Thanks for submitting your paper to JOSS @DavidDoukhan. **Currently, there isn't an JOSS editor assigned** to your paper.
>
> The author's suggestion for the handling editor is @gkthiruvathukal.
>
> @DavidDoukhan if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). In addition, [this list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list).
>
> **Editor instructions**
>
> The JOSS submission bot @editorialbot is here to help you find and assign reviewers and start the main review. To find out what @editorialbot can do for you type:
>
> ```
> @editorialbot commands
> ```

Dear editor, please find below a suggestion of three potential reviewers:

thelinuxmaniac
lukassnoek
vimalthilak

Please let me know il you need any additional informatoin,
