### 2022-02-24 #4195 @arfon
@DavidDoukhan – while I look for an editor, could you please take a look at the DOI suggestions that @editorialbot has made here and if they look correct update your BibTeX file accordingly?
