+++ README.md
@@ -14,12 +14,12 @@ Results can be exported as tables, augmented video streams, or rich ASS subtitle
 The trained age and gender classification model provided is based on a `ResNet50` architecture.
 Evaluation results are highly competitive with respect to the current state-of-the-art, and appear to reduce gender, age and racial biases.
 
-Should you need further details regarding this work, please refer to the following [paper](https://github.com/ina-foss/inaFaceAnalyzer/blob/master/paper.md):
+Should you need further details regarding this work, please refer to the following [paper](https://github.com/openjournals/joss-papers/blob/joss.04195/joss.04195/10.21105.joss.04195.pdf):
 
 ```bibtex
 @journal{doukhan2022joss,
   author = {David Doukhan and Thomas Petit},
-  title = {inaFaceAnalyzer: a Python toolbox for large-scale face-based description of gender representation in media},
+  title = {inaFaceAnalyzer: a Python toolbox for large-scale face-based description of gender representation in media with limited gender, racial and age biases},
   journal = {JOSS - The journal of Open Source Software (currently being reviewed)},
   year = {submission in progress}
 }
