+++ paper.bib
@@ -14,7 +14,8 @@
   number={2},
   pages={144--157},
   year={2018},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s11263-016-0940-3}
 }
 
 
@@ -189,7 +190,8 @@ journal={La revue des médias}
   volume={12},
   number={1},
   pages={68--71},
-  year={2021}
+  year={2021},
+  doi={10.1109/cvpr.2017.53},
 }
 
 
@@ -197,7 +199,7 @@ journal={La revue des médias}
 title={Le Genre dans les imaginaires de l’informatique : voix et représentations dans les publicités (1980-2010)},
 author={Nolwenn Baudry and Camille Claverie and Dylan Nibert and Elsa Poupardin and Antonin Segault and Emma Vadillo Quesada},
 year={2020},
-booktitle={Billets de recherche Inatheque}
+journal={Billets de recherche Inatheque}
 }
 
 @inproceedings{calvez20,
@@ -216,6 +218,7 @@ year={2020},
  year={2021},
  volume={30},
  pages={5032-5044},
+ doi={10.1109/tip.2021.3077144}
  }
 
 
@@ -224,5 +227,6 @@ year={2020},
   author={Danelljan, Martin and H{\"a}ger, Gustav and Khan, Fahad and Felsberg, Michael},
   booktitle={British Machine Vision Conference, Nottingham},
   year={2014},
-  organization={Bmva Press}
+  organization={Bmva Press},
+  doi={10.5244/c.28.65},
 }
