+++ paper.bib
@@ -3,7 +3,8 @@
       author={Karkkainen, Kimmo and Joo, Jungseock},
       booktitle={Proceedings of the IEEE/CVF Winter Conference on Applications of Computer Vision},
       year={2021},
-      pages={1548--1558}
+      pages={1548--1558},
+      doi={10.1109/wacv48630.2021.00159},
 }
 
 @article{rothe2018deep,
@@ -27,7 +28,8 @@
   number={10},
   pages={1090--1104},
   year={2000},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/34.879790},
 }
 
 @inproceedings{serengil2020lightface,
@@ -36,7 +38,8 @@
   booktitle={2020 Innovations in Intelligent Systems and Applications Conference (ASYU)},
   pages={23-27},
   year={2020},
-  organization={IEEE}
+  organization={IEEE},
+  doi={10.1109/asyu50717.2020.9259802},
 }
 
 @inproceedings{he2016deep,
@@ -44,15 +47,19 @@
   author={He, Kaiming and Zhang, Xiangyu and Ren, Shaoqing and Sun, Jian},
   booktitle={Proceedings of the IEEE conference on computer vision and pattern recognition},
   pages={770--778},
-  year={2016}
+  year={2016},
+  doi={10.1109/cvpr.2016.90},
 }
 
-@misc{chollet2015keras,
+
+@online{chollet2015keras,
   title={Keras},
-  author={Chollet, Fran\c{c}ois and others},
+  author={Chollet, Francois and others},
   year={2015},
-  howpublished={\url{https://keras.io}},
+  publisher={GitHub},
+  url={https://github.com/fchollet/keras},
 }
+
 
 @misc{kvggface,
 title={keras-vggface},
@@ -76,7 +83,8 @@ howpublished={\url{https://github.com/rcmalli/keras-vggface}},
   booktitle={CVPR 2011},
   pages={529--534},
   year={2011},
-  organization={IEEE}
+  organization={IEEE},
+  doi={10.1109/cvpr.2011.5995566},
 }
 
 @article{doukhan2018describing,
@@ -107,7 +115,8 @@ howpublished={\url{https://github.com/rcmalli/keras-vggface}},
   author={Zhang, Zhifei, Song, Yang, and Qi, Hairong},
   booktitle={IEEE Conference on Computer Vision and Pattern Recognition (CVPR)},
   year={2017},
-  organization={IEEE}
+  organization={IEEE},
+  doi={10.1109/cvpr.2017.463},
 }
 
 
@@ -119,7 +128,8 @@ howpublished={\url{https://github.com/rcmalli/keras-vggface}},
   number={1},
   pages={1--9},
   year={2021},
-  publisher={Palgrave}
+  publisher={Palgrave},
+  doi={10.1057/s41599-021-00815-9},
 }
 
 @inproceedings{csa2021,
@@ -181,6 +191,7 @@ journal={La revue des médias}
   booktitle={Proceedings of the IEEE Conference on Computer Vision and Pattern Recognition},
   pages={2682--2690},
   year={2017},
+  doi={10.1109/cvpr.2017.53},
 }
 
 @inproceedings{islam2021gender,
