### 2022-02-25 #4195 @DavidDoukhan
> @DavidDoukhan – while I look for an editor, could you please take a look at the DOI suggestions that @editorialbot has made here and if they look correct update your BibTeX file accordingly?

@arfon , I updated the bibtex and it seems better now,

Kind regards,
