### 2022-02-27 976c4a1
worker code improved and refactored
