### 2022-02-27 6b5b04a
default models embed in dockerimage
