### 2022-02-27 5074933
minor modif allowing to have clean version number in docker images
