### 2022-02-27 #4195 @arfon
:wave: @thelinuxmaniac @lukassnoek @vimalthilak, would you be willing to review this submission for JOSS?
