### 2022-03-07 ina-foss/inaFaceAnalyzer#26 @lukassnoek
Author contributions
Hi, I'm reviewing the package for JOSS (see [here](https://github.com/openjournals/joss-reviews/issues/4210)) and I'll create separate issues for the things I encounter during the review process.

Considering the author contributions: @DavidDoukhan clearly did most of the work, but I can't find any contributions in terms of code/documentation of Thomas Petit. In contrast, as is mentioned under 'CREDITS' in the readme, @ZohraRezgui was responsible for the prototype of inaFaceAnalyzer, but is not listed as a co-author. Could the decision to include Thomas Petit but exclude Zohra Rezgui be explained?
