### 2022-03-07 ina-foss/inaFaceAnalyzer#27 @lukassnoek
'Statement of need' in documentation
As part of the [JOSS review](https://github.com/openjournals/joss-reviews/issues/4210):

As outlined in the reviewer instructions, the documentation (in this case: the readme) should include a "A statement of need". Perhaps you could include a short paragraph (copied or paraphrased from the manuscript)?
