### 2022-03-07 ina-foss/inaFaceGender#28 @lukassnoek
Python API documentation
As part of the [JOSS review](https://github.com/openjournals/joss-reviews/issues/4210):

Although I really appreciate the useful tutorial notebooks, I think they do not replace a complete API documentation of the Python interface. Adding a documentation of at least the most important package functionality (e.g., the `VideoAnalyzer`, `VideoKeyframes`, `VideoTracking`,  and `ImageAnalyzer` classes) would be great!
