### 2022-03-07 ina-foss/inaFaceAnalyzer#29 @lukassnoek
Community guidelines
As part of the [JOSS review](https://github.com/openjournals/joss-reviews/issues/4210):

As far as I can see, no community guidelines on how to contribute/report issues/seek help are listed. A short paragraph on how to do this (through github issues/PRs, I assume) would suffice, I think.
