### 2022-03-07 #4210 @lukassnoek
Hi all,

I completed (the first round of) my review for the `inaFaceAnalyzer` package. I think it is a well-developed, feature-complete package with an intuitive and extensive API. Installation went smoothly (tested on a remote server running Ubuntu 20.04 with a Quadro RTX 8000 GPU; installed without a conda environment) and all outlined functionality worked as expected. In addition, the tutorial notebooks were very helpful. Because of the remote server, I could not test all functionality (e.g., webcam based tracking, distributed analyses, etc.).

I encountered several issues that I think need to be addressed, which were submitted as issues in the github repository of the package (see mentions above this comment). I hope this will improve the (documentation of) the package!
