### 2022-03-09 ina-foss/inaFaceAnalyzer#26 @DavidDoukhan
Dear @lukassnoek ,

I proposed @ZohraRezgui to be part of the author list and/or to contribute to our Joss paper proposal. I believe her work is a significant contribution to this software and as such - including Zohra as an author of this submission was my wish.

Zohra told me she does not want to be part of the co-author list of this study and I respect her choice. If she changes her mind after reading this thread, we’ll be very happy to include her to the author list of our proposal. Consequently, we displayed her contributions in the paper’s acknowledgement section and in github’s repository CREDITS section.

You’ll find more details below. Please let us know if you have suggestions of changes to be done in the paper and/or in github’s repository in order to make this information more explicit.

Zohra Rezgui did a 6 months internship at INA in 2019 under my supervision. References to her internship report are provided in JOSS paper and on github repository. During her internship, she implemented a face processing pipeline, and trained face gender classification models.

Thomas Petit is a PhD candidate at INA working on large-scale face recognition. He provided us several pieces of code he realized during his PhD: face detection, face alignment, facial landmark extraction, facial pre-trained neural representation models. He also contributed to Zohra’s supervision by providing several valuable advices. I also benefited from his expertise in face recognition and image processing while working on this software.

At the end of Zohra’s internship, I asked Zohra to commit on Github the prototype she realized during her internship in order to make her contribution to this project explicit. This prototype includes some adaptations of Thomas Petit’s code that were committed to the repository as Zohra. Thomas Petit made explicit contributions to the repository with 2 github avatars: @tpetit and @w2ex . While these contributions can be seen in the project network graph, Thomas’s Github identifier is not displayed in github repository’s contributor list.

Zohra’s face gender classification model performances are not described in our paper proposal but are described in her internship report. The newly created models provided in inaFaceAnalyzer outperform the models she realized during her internship, for all evaluation metrics and corpora. Given the short format required for Joss journal I chose to focus on the performance of the model we recommend to use, rather than on older models. Zohra’s model is kept in the distribution for historical and reproducibility reasons.

As you noticed, I made several major contributions to this software: training and evaluation of new models, debugging, profiling, refactoring, packaging, doc, tests, command line scripts, etc…
