### 2022-03-11 ina-foss/inaFaceAnalyzer#26 @lukassnoek
Hi @DavidDoukhan, thank you for your clarification. Let's wait for Zohra, but if we don't hear anything by the time all the other issues are resolved, I'll consider this issue closed.
