+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -7,32 +7,39 @@ assignees: ''
 
 ---
 
-**Describe the bug**
-<!-- A clear and concise description of what the bug is. -->
+## Describe the bug
+A clear and concise description of what the bug is.
 
-**To Reproduce**
+
+## To Reproduce
  - provide the lines of code that produce an unexpected behavior or the command line program used together with the commandline arguments
  - downloadable link to the input data associated (images, videos) to the unexpected behavior
 
-**Expected behavior**
+## Expected behavior
+
 A clear and concise description of what you expected to happen.
 
-**Current behavior**
+## Current behavior
+
 Tell us what happens instead of the expected behavior.
 Include any log/backtrace here.
 
 
-**System information**
- - OS and distribution (eg: Linux Ubuntu 20.04)
- - Python version
- - Tensorflow version
- - output of command `pip freeze`
- - Amount of available RAM
- - Running on GPU or CPU
- - output of command `nvidia-smi` (if GPU is used)
- - CUDA and CuDNN versions (if GPU is used)
- - installation type (from sources, from pypi, from dockerhub ?)
- - inaFaceAnalyzer version
+## Environment
+Please provide the following details on your environment : 
 
-**Additional context**
+
+ - OS and distribution (eg: Linux Ubuntu 20.04) : 
+ - Python version :
+ - Tensorflow version :
+ - output of command `pip freeze` :
+ - Amount of available RAM :
+ - Running on GPU or CPU :
+ - output of command `nvidia-smi` (if GPU is used) :
+ - CUDA and CuDNN versions (if GPU is used) :
+ - installation type (from sources, from pypi, from dockerhub ?) :
+ - inaFaceAnalyzer version :
+
+## Additional context
+
 Add any other context about the problem here.
