+++ .github/CONTRIBUTING.md
@@ -3,7 +3,7 @@
 Those are the main contributing guidelines for contributing to this project:
 
 - Verify that your contribution does not embark proprietary code or infringe any copyright of any sort.
-- Avoid adding any unnecessary dependencies to the project, espcially of those are not easily packaged and installed through `conda` or `pip`.
+- Avoid adding any unnecessary dependencies to the project, espcially of those are not easily packaged and installed through `pip`.
 - Python contributions must follow the [PEP 8 style guide](https://www.python.org/dev/peps/pep-0008/).
 - Use [Pull Request](https://help.github.com/en/github/collaborating-with-issues-and-pull-requests/about-pull-requests) mechanism and please be patient while waiting for reviews.
 - Remain polite and civil in all exchanges with the maintainers and other contributors.
