+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -1,0 +1,16 @@
+# Pull request title
+
+- [ ] I read [contributing guideline](https://github.com/ina-foss/inaFaceAnalyzer/blob/master/.github/CONTRIBUTING.md)
+- [ ] I didn't find a similar pull request already open.
+
+## Description
+
+A few sentences describing the overall goals of the pull request's commits.
+
+## How this contribution was tested
+
+- [ ] I implemented unit tests which ran successfully using `python test_inaFaceAnalyzer.py`
+
+## Documentation link and external references
+
+Please provide any info that may help us better understand your code.
