### 2022-03-11 ina-foss/inaFaceAnalyzer#29 @DavidDoukhan
Dear @lukassnoek ,

Thanks for this issue : community guidelines were clearly missing and are now improved !

Please let me know if you think some additinal information will help to improve adoption and encourage software contributions to `inaFaceAnalyzer`.

Few changes were done according to your remark :
* few words have been said at the end of  README.md in the contributing sections
* contributing guidelines were added in .github/CONTRIBUTING.md
* a pull request template has been added in .github/PULL_REQUEST_TEMPLATE.md
* the bug report template has been updated in .github/ISSUE_TEMPLATE/bug_report.md
