### 2022-03-11 ina-foss/inaFaceAnalyzer#27 @DavidDoukhan
Dear @lukassnoek ,

I've just added a "statement of need" section in the README.md
This also describes the latest use of the software in a French public report.

Please let me know if you think additional improvements could be done to this section.

Kind regards,
