### 2022-03-21 ina-foss/inaFaceAnalyzer#27 @lukassnoek
Looks great! Closing this issue.
