### 2022-03-21 ina-foss/inaFaceAnalyzer#29 @lukassnoek
Much improved indeed. Closing this issue.
