### 2022-03-21 ina-foss/inaFaceAnalyzer#26 @lukassnoek
Given that we haven't heard from Zohra, I'm considering this issue resolved.
