### 2022-04-07 #4210 @lukassnoek
Hi @arfon , sounds about right! @DavidDoukhan already addressed some of the issues I raised. I think only one is remaining (documentation of Python API).
