+++ docs/index.rst
@@ -1,0 +1,35 @@
+.. inaFaceAnalyzer documentation master file, created by
+   sphinx-quickstart on Wed May 18 14:58:19 2022.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+Welcome to inaFaceAnalyzer's documentation!
+===========================================
+
+
+API documentation
+=================
+
+.. toctree::
+   :maxdepth: 2
+
+   classifier
+
+
+
+
+
+
+
+
+.. include :: authors.rst
+
+
+
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
