+++ docs/classifier.rst
@@ -1,0 +1,9 @@
+Choosing a face classification model
+------------------------------------
+
+.. automodule:: inaFaceAnalyzer.face_classifier
+
+Abstract Face Classifier
+^^^^^^^^^^^^^^^^^^^^^^^^
+
+.. autoclass:: inaFaceAnalyzer.face_classifier.FaceClassifier
