+++ docs/analyzer.rst
@@ -1,0 +1,26 @@
+Media Analyzer Classes
+----------------------
+
+.. automodule:: inaFaceAnalyzer.inaFaceAnalyzer
+
+Abstract Media Analyzer Class 
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+.. autoclass:: inaFaceAnalyzer.inaFaceAnalyzer.FaceAnalyzer
+
+ImageAnalyzer
+^^^^^^^^^^^^^
+
+
+.. autoclass:: inaFaceAnalyzer.inaFaceAnalyzer.ImageAnalyzer
+    :members:
+    :special-members:
+
+.. autoclass:: inaFaceAnalyzer.inaFaceAnalyzer.VideoAnalyzer
+
+.. autoclass:: inaFaceAnalyzer.inaFaceAnalyzer.VideoKeyframes
+
+.. autoclass:: inaFaceAnalyzer.inaFaceAnalyzer.VideoTracking
+
+
+
