+++ docs/classifier.rst
@@ -7,3 +7,15 @@ Abstract Face Classifier
 ^^^^^^^^^^^^^^^^^^^^^^^^
 
 .. autoclass:: inaFaceAnalyzer.face_classifier.FaceClassifier
+
+.. autoclass:: inaFaceAnalyzer.face_classifier.Resnet50FairFaceGRA
+   :members:
+   :special-members:
+
+.. autoclass:: inaFaceAnalyzer.face_classifier.OxfordVggFace
+   :members:
+   :special-members:
+
+.. autoclass:: inaFaceAnalyzer.face_classifier.Vggface_LSVM_YTF
+   :members:
+   :special-members:
