+++ docs/requirements.readthedocs.txt
@@ -1,0 +1,1 @@
+inaFaceAnalyzer
