+++ docs/requirements.readthedocs.txt
@@ -1,1 +1,2 @@
+cmake
 inaFaceAnalyzer
