+++ docs/requirements.readthedocs.txt
@@ -1,2 +1,0 @@
-cmake
-inaFaceAnalyzer
