+++ .readthedocs.yaml
@@ -28,7 +28,6 @@ formats: all
 
 # Optionally declare the Python requirements required to build your docs
 python:
-  version: "3.7"
   install:
     - method: pip
       path: .
