+++ .readthedocs.yaml
@@ -20,8 +20,6 @@ build:
 # Build documentation in the docs/ directory with Sphinx
 sphinx:
    configuration: docs/conf.py
-   fail_on_warning: true
-
 
 # If using Sphinx, optionally build your docs in additional formats such as PDF
 formats: all
