+++ docs/classifier.rst
@@ -3,19 +3,23 @@ Choosing a face classification model
 
 .. automodule:: inaFaceAnalyzer.face_classifier
 
-Abstract Face Classifier
-^^^^^^^^^^^^^^^^^^^^^^^^
 
 .. autoclass:: inaFaceAnalyzer.face_classifier.FaceClassifier
+    :members: preprocessed_img_list
+    :special-members: __call__
+
+.. autoclass:: inaFaceAnalyzer.face_classifier.Resnet50FairFace
+   :show-inheritance:
 
 .. autoclass:: inaFaceAnalyzer.face_classifier.Resnet50FairFaceGRA
-   :members:
-   :special-members:
+   :show-inheritance:
 
 .. autoclass:: inaFaceAnalyzer.face_classifier.OxfordVggFace
-   :members:
-   :special-members:
+   :show-inheritance:
 
 .. autoclass:: inaFaceAnalyzer.face_classifier.Vggface_LSVM_YTF
-   :members:
-   :special-members:
+   :show-inheritance:
+
+
+.. autoclass:: inaFaceAnalyzer.face_classifier.Vggface_LSVM_FairFace
+   :show-inheritance:
