### 2022-05-20 c0e9e7c
_static warning removed
