### 2022-05-20 e176930
minor refactoring in classifiers - do not return intermediate features any more documentation in progress
