+++ docs/conf.py
@@ -27,7 +27,10 @@ author = 'David Doukhan'
 # Add any Sphinx extension module names here, as strings. They can be
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
-extensions = ['sphinx.ext.autodoc', 'sphinx.ext.viewcode', 'sphinx.ext.napoleon']
+extensions = ['sphinx.ext.autodoc',
+              'sphinx.ext.viewcode',
+              'sphinx.ext.napoleon',
+              'sphinx_toolbox.more_autodoc.autonamedtuple']
 
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ['_templates']
