### 2022-05-20 cbf615a
minor refactoring of libfacedetection -> do not return eye position anymore: thids feature was not used
doc in progress
