+++ media/README.txt
@@ -6,7 +6,7 @@ From https://www.flickr.com/photos/9518198@N04/1546923312
 Europa21_-_2.jpg and diallo224.jpg
 From https://commons.wikimedia.org/wiki/File:Europa21_-_2.jpg
 
-20091020222328!KnuthAtOpenContentAlliance.jpg and knuth224.jpg
+dknuth.jpg and knuth224.jpg
 From https://commons.wikimedia.org/wiki/File:KnuthAtOpenContentAlliance.jpg
 
 
