### 2022-05-20 299aa48
doc update + default documentation instructions in conf.py
