+++ docs/data_structures.rst
@@ -1,0 +1,14 @@
+Data Structures
+---------------
+
+Blah Blah Blah
+autrement pandas dataframes
+
+.. autonamedtuple:: inaFaceAnalyzer.rect.Rect
+    :members:
+
+.. autonamedtuple:: inaFaceAnalyzer.face_detector.Detection
+    :members:
+
+.. autonamedtuple:: inaFaceAnalyzer.face_tracking.TrackDetection
+    :members:
