+++ docs/detector.rst
@@ -3,25 +3,19 @@ Defining a face detection method
 
 .. automodule:: inaFaceAnalyzer.face_detector
 
-.. autonamedtuple:: inaFaceAnalyzer.face_detector.Detection
-    :members:
-
 .. autoclass:: inaFaceAnalyzer.face_detector.FaceDetector
     :members:
-    :special-members:
+    :special-members: __init__,__call__
 
 .. autoclass:: inaFaceAnalyzer.face_detector.LibFaceDetection
-    :show-inheritance:
     :members:
     :special-members:
 
 
 .. autoclass:: inaFaceAnalyzer.face_detector.OcvCnnFacedetector
-    :show-inheritance:
     :members:
     :special-members:
 
 .. autoclass:: inaFaceAnalyzer.face_detector.IdentityFaceDetector
-    :show-inheritance:
     :members:
     :special-members:
