### 2022-05-20 1ac1267
comments + minor refactoring
