+++ .readthedocs.yaml
@@ -28,7 +28,7 @@ formats: all
 python:
   install:
     - method: pip
-      path: sphinx_toolbox
-    - method: pip
       path: .
-#   - requirements: docs/requirements.txt
+      extra_requirements:
+        - sphinx_toolbox
+
