### 2022-05-21 afeb269
extra dependency for building docs
