+++ .readthedocs.yaml
@@ -30,5 +30,4 @@ python:
     - method: pip
       path: .
       extra_requirements:
-        - sphinx_toolbox
-
+        - doc
