+++ inaFaceAnalyzer/rect.py
@@ -33,88 +33,39 @@ class Rect(NamedTuple):
     """
 
     #: left
-    x1 : float # left
+    x1 : float
     #: top
-    y1 : float # top
+    y1 : float
     #: right
-    x2 : float # right
+    x2 : float
     #: bottom
-    y2 : float # bottom
+    y2 : float
 
     @property
     def w(self):
-        """ Rect width """
+        """ self width """
         return self.x2 - self.x1
 
     @property
     def h(self):
-        """ Rect height """
+        """ self height """
         return self.y2 - self.y1
 
     @property
     def center(self):
-        """ Center (x,y) of Rect (x1, y1, x2, y2) """
+        """ center (x,y) of self (x1, y1, x2, y2) """
         x1, y1, x2, y2 = self
         return ((x1 + x2) / 2), ((y1 + y2) / 2)
 
     @property
     def area(self):
-        """ Surface area """
+        """ self Surface area """
         return self.w * self.h
 
     @property
     def max_dim_len(self):
         """ max (width, height)"""
         return max(self.h, self.w)
-
-    def transpose(self, xoffset, yoffset):
-        """
-        Translation
-        Args:
-            xoffset (float): horizontal offset.
-            yoffset (float): vertical offset.
-        Returns:
-            Rect : translated Rect
-        """
-        x1, y1, x2, y2 = self
-        return Rect(x1 + xoffset, y1 + yoffset, x2 + xoffset, y2 + yoffset)
-
-    def mult(self, x, y):
-        x1, y1, x2, y2 = self
-        return Rect(x1 * x, y1 * y, x2 * x, y2 * y)
-
-    def intersect(self, r):
-        x1, y1, x2, y2 = self
-        ret = Rect(max(x1, r.x1), max(y1, r.y1), min(x2, r.x2), min(y2, r.y2))
-        if ret.h <= 0 or ret.w <= 0:
-            return Rect(0,0,0,0)
-        return ret
-
-    def iou(self, r):
-        inter = self.intersect(r).area
-        union = self.area + r.area - inter
-        return inter / union
-
-    def __contains__(self, point):
-        x1, y1, x2, y2 = self
-        x, y = point
-        return x >= x1 and x <= x2 and y >= y1 and y <= y2
-
-    def to_int(self):
-        return Rect(*[int(round(e)) for e in self])
-
-    @staticmethod
-    def from_dlib(x):
-        """ Import to internal data structure dlib's native rectangle"""
-        return Rect(x.left(), x.top(), x.right(), x.bottom())
-
-    def to_dlibInt(self):
-        """ Convert internal data structure to dlib's int rectangle"""
-        return dlib.rectangle(*[e for e in self.to_int()])
-
-    def to_dlibFloat(self):
-        """ Convert internal data structure to dlib's float rectangle"""
-        return dlib.drectangle(*self)
 
     @property
     def square(self):
@@ -123,8 +74,139 @@ class Rect(NamedTuple):
         xc, yc = self.center
         return Rect(xc - offset, yc - offset, xc + offset, yc + offset)
 
+
+    def transpose(self, xoffset, yoffset):
+        """
+        Translation
+
+        Args:
+            xoffset (float): horizontal offset.
+            yoffset (float): vertical offset.
+        Returns:
+            Rect :
+        """
+        x1, y1, x2, y2 = self
+        return Rect(x1 + xoffset, y1 + yoffset, x2 + xoffset, y2 + yoffset)
+
+    def mult(self, x, y):
+        """
+        Multiply self coordinates by horizontal and vertical scaling factors
+        Usefull for converting [0...1] coordinates to image frame dimensions in pixels
+
+        Args:
+            x (float: horizontal scaling factor.
[…]
