### 2022-05-21 9b6a31f
minor refactoring display doc done
