+++ docs/classifier.rst
@@ -3,23 +3,17 @@ Choosing a face classification model
 
 .. automodule:: inaFaceAnalyzer.face_classifier
 
-
 .. autoclass:: inaFaceAnalyzer.face_classifier.FaceClassifier
     :members: preprocessed_img_list
     :special-members: __call__
 
 .. autoclass:: inaFaceAnalyzer.face_classifier.Resnet50FairFace
-   :show-inheritance:
 
 .. autoclass:: inaFaceAnalyzer.face_classifier.Resnet50FairFaceGRA
-   :show-inheritance:
 
 .. autoclass:: inaFaceAnalyzer.face_classifier.OxfordVggFace
-   :show-inheritance:
 
 .. autoclass:: inaFaceAnalyzer.face_classifier.Vggface_LSVM_YTF
-   :show-inheritance:
-
 
 .. autoclass:: inaFaceAnalyzer.face_classifier.Vggface_LSVM_FairFace
-   :show-inheritance:
+
