+++ docs/authors.rst
@@ -1,6 +1,0 @@
-Authors
-=======
-* David Doukhan <ddoukhan@ina.fr>
-* Zohra Rezgui <zohra.rzg@gmail.com>
-* Thomas Petit : tpetit@ina.fr
-
