+++ docs/index.rst
@@ -6,13 +6,24 @@
 Welcome to inaFaceAnalyzer's documentation!
 ===========================================
 
+inaFaceAnalyzer is a Python toolbox designed for large-scale analysis of faces in image or video streams.
+It provides a modular processing pipeline allowing to predict age and gender from faces.
 
-API documentation
-=================
+Should you need further details regarding this work, please refer to the following
+`paper <https://github.com/openjournals/joss-papers/blob/joss.04210/joss.04210/10.21105.joss.04210.pdf>`_:
+
+David Doukhan and Thomas Petit (2022),
+inaFaceAnalyzer: a Python toolbox for large-scale face-based description of gender representation in media with limited gender, racial and age biases,
+*JOSS - The journal of Open Source Software (currently being reviewed)*
+
+
+User Guide and API documentation
+================================
 
 .. toctree::
    :maxdepth: 2
 
+   install
    analyzer
    detector
    classifier
