+++ docs/index.rst
@@ -11,7 +11,6 @@ It provides a modular processing pipeline allowing to predict age and gender fro
 
 Should you need further details regarding this work, please refer to the following
 `paper <https://github.com/openjournals/joss-papers/blob/joss.04210/joss.04210/10.21105.joss.04210.pdf>`_:
-
 David Doukhan and Thomas Petit (2022),
 inaFaceAnalyzer: a Python toolbox for large-scale face-based description of gender representation in media with limited gender, racial and age biases,
 *JOSS - The journal of Open Source Software (currently being reviewed)*
@@ -24,6 +23,13 @@ User Guide and API documentation
    :maxdepth: 2
 
    install
+
+API documentation
+=================
+
+.. toctree::
+   :maxdepth: 2
+
    analyzer
    detector
    classifier
