+++ docs/classifier.rst
@@ -4,10 +4,12 @@ Choosing a face classification model
 .. automodule:: inaFaceAnalyzer.face_classifier
 
 .. autoclass:: inaFaceAnalyzer.face_classifier.FaceClassifier
-    :members: preprocessed_img_list
+    :members: preprocessed_img_list, input_shape,bbox_scale, bbox2square
     :special-members: __call__
 
 .. autoclass:: inaFaceAnalyzer.face_classifier.Resnet50FairFace
+    :members: input_shape, bbox_scale, bbox2square
+    :inherited-members:
 
 .. autoclass:: inaFaceAnalyzer.face_classifier.Resnet50FairFaceGRA
 
