+++ inaFaceAnalyzer/face_detector.py
@@ -26,7 +26,9 @@
 """
 Face detection classes are in charge of finding faces in image frames.
 
-Two face detection classes are provided : :class:`LibFaceDetection` (default) and :class:`OcvCnnFacedetector`.
+Two face detection classes are provided : 
+    - :class:`LibFaceDetection` (default) 
+    - :class:`OcvCnnFacedetector`.
 
 
 Face detection classes inherits from abstract class :class:`FaceDetector` and share a common interface.
@@ -68,6 +70,7 @@ class Detection(NamedTuple):
     #: face detection confidence (0 = lowest confidence, 1 = highest confidence)
     detect_conf : float
 
+
 # Currently, we wish to use the same eye detection procedure for all detection engines
 # class DetectionEyes(NamedTuple):
 #     """
@@ -80,6 +83,13 @@ class Detection(NamedTuple):
 
 
 class FaceDetector(ABC):
+    
+    #    @classmethod
+    #    @abstractmethod
+    #    def output_type() : pass
+    output_type = Detection
+
+    
     def __init__(self, minconf, min_size_px, min_size_prct, padd_prct):
         """
         Common face detection constructor
@@ -97,7 +107,17 @@ class FaceDetector(ABC):
         self.padd_prct = padd_prct
 
     def __call__(self, frame, verbose=False):
+        """
+        Perform face detection on image frames
 
+        Args:
+            frame (:class:`numpy.ndarray`): RGB image frame (height, width, 3).
+            verbose (bool, optional): display intermediate results such as detected faces Not to be used in production. Defaults to False.
+
+        Returns:
+            list of :class:`Detection` instances
+        """        
+        
         tmpframe = frame
 
         if self.padd_prct:
@@ -132,21 +152,20 @@ class FaceDetector(ABC):
 
         return lret
 
-#    @classmethod
-#    @abstractmethod
-#    def output_type() : pass
-    output_type = Detection
 
     @abstractmethod
     def _call_imp(self, frame): pass
 
     def most_central_face(self, frame, contain_center=True, verbose=False):
         """
+        To be used for processing ML datasets and training new face classification models.
+        
+        Some ML face corpora images containing several faces, with the target annotated face at the center.
+        
         This method returns the detected face which is closest from the center of the image frame
-        Usefull for preprocessing ML face datasets containing several faces per image
 
         Args:
-            frame (numpy.ndarray (height, width, 3)): RGB image data.
+            frame (:class:`numpy.ndarray`): RGB image frame (height, width, 3)
             contain_center (bool, optional): if True, the returned face MUST include image center. Defaults to True.
             verbose (bool, optional): Display detected faces. Defaults to False.
 
@@ -173,13 +192,26 @@ class FaceDetector(ABC):
 
     def get_closest_face(self, frame, ref_bbox, min_iou=.7, squarify=True, verbose=False):
         """
-        Some face corpora may contain pictures with several faces
-        together with the reference bounding box of annotated faces
-        This function is aimed at preprocessing such face corpora
-        Automatic face detection is used, and the detected face with the largest
-        iou with the reference bounding box is returned
-        if no detected face corresponds to this IOU criteria, returns None
+        To be used for processing ML datasets and training new face classification models.
+    
+        Some face corpora images may contain several annotated faces.
+        This method return the detected face having the largest IOU with target ref_box.
+        The IOU must be > to min_iou.
+
+        Args:
+            frame (:class:`numpy.ndarray`): RGB image frame (height, width, 3).
+            ref_bbox (tuple or Rect): reference face bounding box (x1, y1, x2, y2).
[…]
