+++ docs/analyzer.rst
@@ -1,29 +1,26 @@
-Media Analyzer Classes
-----------------------
+Media processing engines
+------------------------
 
 .. automodule:: inaFaceAnalyzer.inaFaceAnalyzer
 
-Abstract Media Analyzer Class
-^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
 .. autoclass:: inaFaceAnalyzer.inaFaceAnalyzer.FaceAnalyzer
     :members:
     :special-members:
 
-ImageAnalyzer
-^^^^^^^^^^^^^
 
 
 .. autoclass:: inaFaceAnalyzer.inaFaceAnalyzer.ImageAnalyzer
     :members:
     :special-members: __init__,__call__
-    :inherited-members: __init__
-
+ 
 .. autoclass:: inaFaceAnalyzer.inaFaceAnalyzer.VideoAnalyzer
-
+    :special-members: __init__,__call__
+ 
 .. autoclass:: inaFaceAnalyzer.inaFaceAnalyzer.VideoKeyframes
+    :special-members: __init__,__call__
 
 .. autoclass:: inaFaceAnalyzer.inaFaceAnalyzer.VideoTracking
-
+    :special-members: __init__,__call__
 
 
