### 2022-06-11 ina-foss/inaFaceAnalyzer#31 @vimalthilak
Exception thrown when I run ina_face_analyzer.py on a single image (Not a face image)
Make sure you ran the unit tests before submitting a bug report and tell us if and where they fail.
You can use the program `test_inaFaceAnalyzer.py` at the root of this repository.

## Describe the bug
A clear and concise description of what the bug is.

## To Reproduce
 - ina_face_analyzer.py --engine image -i <non-face-image> -o results.csv
 - Note that the image I used to test is a PDF file of a Matplotlib plot. I can't share the exact image but I hope the description can help you generate a test image for your use

## Expected behavior
File should be processed without throwing an exception

## Current behavior

Tell us what happens instead of the expected behavior.
Include any log/backtrace here.
Traceback (most recent call last):
  File "$HOME/miniconda3/envs/face/bin/ina_face_analyzer.py", line 139, in <module>
    df = engine(args.input)
  File "$HOME/miniconda3/envs/face/lib/python3.9/site-packages/inaFaceAnalyzer/inaFaceAnalyzer.py", line 217, in __call__
    return self._process_stream(stream, self.face_detector)
  File "$HOME/miniconda3/envs/face/lib/python3.9/site-packages/inaFaceAnalyzer/inaFaceAnalyzer.py", line 161, in _process_stream
    for iframe, frame in stream_iterator:
  File "$HOME/miniconda3/envs/face/lib/python3.9/site-packages/inaFaceAnalyzer/opencv_utils.py", line 84, in image_iterator
    yield f, imread_rgb(f, verbose)
  File "$HOME/miniconda3/envs/face/lib/python3.9/site-packages/inaFaceAnalyzer/opencv_utils.py", line 128, in imread_rgb
    frame = cv2.cvtColor(img, cv2.COLOR_BGR2RGB)
cv2.error: OpenCV(4.6.0) /io/opencv/modules/imgproc/src/color.cpp:182: error: (-215:Assertion failed) !_src.empty() in function 'cvtColor'

## Environment
Please provide the following details on your environment :

 - OS and distribution: Linux Ubuntu 18.04
 - Python version : 3.9.12
 - Tensorflow version : 2.9.1
 - output of command `pip freeze` :
absl-py==1.1.0
astunparse==1.6.3
av==9.2.0
cachetools==5.2.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
Cheetah3==3.2.6.post1
cycler==0.11.0
dlib==19.24.0
flatbuffers==1.12
fonttools==4.33.3
gast==0.4.0
google-auth==2.7.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.46.3
h5py==3.7.0
idna==3.3
importlib-metadata==4.11.4
inaFaceAnalyzer==0.6.1
joblib==1.1.0
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
libclang==14.0.1
Markdown==3.3.7
matplotlib==3.5.2
numpy==1.22.4
oauthlib==3.2.0
onnxruntime-gpu==1.11.1
opencv-contrib-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.2
Pillow==9.1.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
Pyro4==4.82
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.1.1
scipy==1.8.1
serpent==1.40
six==1.16.0
sklearn==0.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
threadpoolctl==3.1.0
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.0

 - Amount of available RAM :
 - Running on GPU or CPU :
 - output of command `nvidia-smi` (if GPU is used) :
 - CUDA and CuDNN versions (if GPU is used) :
 - installation type (from sources, from pypi, from dockerhub ?) :
 - inaFaceAnalyzer version :

## Additional context

I am following your README as a reviewer for your OSS contribution. See https://github.com/openjournals/joss-reviews/issues/4210
