### 2022-06-13 #4210 @DavidDoukhan
> @arfon and Authors: Thanks for following up and also for your patience. I am done with my initial review of https://github.com/ina-foss/inaFaceAnalyzer/. The package is well written and has many components expected by users ( & reviewers). Two issues
>
> * I noted one issue while doing a test run on my side
> * I didn't see a mention of "related work" in terms of other packages that cover portions of the proposed package. If the Authors can clarify that either here (maybe my misunderstanding/oversight) or in the paper then that'd be great

Dear @vimalthilak , I'm not sure to fully understand what you mean with a mention of "related work" (i'm not native English speaker).

Do you mean a listing of the software components and libraries used to build inaFaceAnalyzer ?
Or do you rather mean a listing of open-source softwares adressing similar problems ?

Kind regards,
