### 2022-06-14 #4210 @vimalthilak
> Dear @vimalthilak , I'm not sure to fully understand what you mean with a mention of "related work" (i'm not native English speaker).

Hi @DavidDoukhan : Thanks for the clarifying question. When I reread my comment I observed that perhaps I could have phrased it better.

>Do you mean a listing of the software components and libraries used to implement inaFaceAnalyzer ?
Or do you rather mean a listing of open-source softwares adressing similar problems ?

I was wondering if you have come across similar open-source of even commercial tools that addresses the problem you are solving with  ```inaFaceAnalyzer```. This information would be helpful in providing context for your work and perhaps show off the areas where ```inaFaceAnalyzer``` shines compared to its peers.

Best,
-Vimal
