### 2022-06-14 ina-foss/inaFaceAnalyzer#31 @vimalthilak
@DavidDoukhan Thanks for the thoughtful and detailed response.
Raising an exception is fine based on your explanation. Does your documentation note that PDF is not a supported input format? If this information is available in the documentation then this issue can be closed.
