### 2022-11-27 #4210 @arfon
@DavidDoukhan – just checking in again here. We've not heard from you since mid-June on this submission. Could you please clarify your intent here – are you still interested in pursing this publication in JOSS?
